Created a small burn algorithm to determine the row-col indices
This is the same type of burn algorithm used in Scalar Segment Features. The algorithm
uses a tolerance of 100 units (probably pixels) to group the origins of the images
into row-col indices. This should allow for some variation in the X,Y value for each
origin and still have it assigned to a consistent row-col index.

Also removed dead code: ZeissDirectoryPatterns.*

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>